+++ src/Controller/CantineController.php
@@ -19,9 +19,11 @@ final class CantineController extends AbstractController
     #[Route('/cantine', name: 'app_cantine')]
     public function index(MenuRepository $menuRepository, ReservationRepository $reservationsRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $menus = $menuRepository->findAll();
         $reservations = $reservationsRepository->findAll();
-        // Convertir les réservations en un tableau de dates formatées
 
 
         return $this->render('cantine/index.html.twig', [
@@ -35,19 +37,20 @@ final class CantineController extends AbstractController
     #[Route('/reserve', name: 'reserve_menu', methods: ['POST'])]
     public function reserveMenu(Request $request, EntityManagerInterface $entityManager, ReservationRepository $reservationRepository): JsonResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $date = $request->request->get('date');
         $user = $this->getUser();
         if (!$user) {
             throw new AccessDeniedException('Vous devez être connecté pour effectuer une réservation.');
         }
 
-        // Vérifie si la date est déjà réservée
         $existingReservation = $reservationRepository->findOneBy(['date' => new \DateTime($date), 'user' => $user]);
         if ($existingReservation) {
             return new JsonResponse(['error' => 'This date is already reserved.'], JsonResponse::HTTP_BAD_REQUEST);
         }
 
-        // Crée une nouvelle réservation
         $reservation = new Reservation();
         $reservation->setUser($user);
         $reservation->setDate(new \DateTime($date));
@@ -59,15 +62,15 @@ final class CantineController extends AbstractController
     }
 
 
-   // CantineController.php
 
 #[Route('/unreserve', name: 'unreserve_menu', methods: ['POST'])]
 public function unreserveMenu(Request $request, EntityManagerInterface $entityManager, ReservationRepository $reservationRepository): JsonResponse
 {
+    if (!$this->getUser()) {
+        return $this->redirectToRoute('app_home_page');
+    }
     $date = $request->request->get('date');
-    $user = $this->getUser(); // Assure-toi que l'utilisateur est connecté
-
-    // Trouve la réservation à supprimer
+    $user = $this->getUser(); 
     $reservation = $reservationRepository->findOneBy(['date' => new \DateTime($date), 'user' => $user]);
     if ($reservation) {
         $entityManager->remove($reservation);
