+++ src/Controller/TrottinetteController.php
@@ -18,6 +18,9 @@ class TrottinetteController extends AbstractController
     #[Route('/trottinettes', name: 'app_trottinette')]
     public function index(Request $request, TrottinetteRepository $trottinetteRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $status = $request->query->get('status');
 
         if ($status) {
@@ -34,6 +37,9 @@ class TrottinetteController extends AbstractController
     #[Route('/trottinettes/ajout', name: 'app_trottinette_ajout')]
     public function ajouter(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         $trottinette = new Trottinette();
         $form = $this->createForm(TrottinetteType::class, $trottinette);
         $form->handleRequest($request);
@@ -54,6 +60,9 @@ class TrottinetteController extends AbstractController
     #[Route('/trottinettes/{id}', name: 'app_trottinette_details', requirements: ['id' => '\d+'])]
     public function details(Trottinette $trottinette): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         return $this->render('trottinette/details.html.twig', [
             'trottinette' => $trottinette,
         ]);
@@ -62,6 +71,9 @@ class TrottinetteController extends AbstractController
     #[Route('/trottinette/{id}/supprimer', name: 'app_trottinette_supprimer')]
     public function supprimer(Trottinette $trottinette, EntityManagerInterface $entityManager): RedirectResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }
         if ($this->getUser()->getType() === 'Administration' || $this->getUser()->getNiveau() === 3) {
             $entityManager->remove($trottinette);
             $entityManager->flush();
@@ -77,9 +89,9 @@ class TrottinetteController extends AbstractController
     #[Route('/trottinette/{id}/reserver', name: 'app_trottinette_reserver')]
     public function reserver(Trottinette $trottinette, EntityManagerInterface $entityManager): RedirectResponse
     {
-
-        // Récupérer l'utilisateur connecté
-        $user = $this->getUser();
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_home_page');
+        }        $user = $this->getUser();
         if (!$user) {
             $this->addFlash('error', 'Vous devez être connecté pour réserver une trottinette.');
             return $this->redirectToRoute('app_login');
@@ -94,27 +106,24 @@ class TrottinetteController extends AbstractController
         $reservation->setUtilisateur($this->getUser());
         $reservation->setDateReservation(new \DateTime());
 
-        // Incrémenter le niveau de l'utilisateur
         $user->setPoints($user->getPoints() + 1);
 
-        // Vérifier si le niveau doit être mis à jour
         if ($user->getPoints() == 30) {
             $user->setNiveau(2);
         } elseif ($user->getPoints() == 80) {
             $user->setNiveau(3);
         }
 
-        // Mettre à jour la session
         $session = $request->getSession();
         $userData = $session->get('user_data', []);
-        $userData['points'] = $user->getPoints(); // Mettre à jour avec la nouvelle valeur
+        $userData['points'] = $user->getPoints(); 
         $userData['niveau'] = $user->getNiveau();
         $session->set('user_data', $userData);
 
         $trottinette->setStatut('Indisponible');
 
         $entityManager->persist($reservation);
-        $entityManager->persist($user); // Mettre à jour l'utilisateur
+        $entityManager->persist($user); 
         $entityManager->persist($trottinette);
         $entityManager->flush();
